Rename sendHttpRequest() for clarity
